requirements.txt
# ===== REQUIRED (Original Project) =====
# Required for core extraction + indexing pipeline.
langchain-text-splitters
faiss-cpu
sentence-transformers
pypdf
numpy

# ===== ADDITIONAL (OCR Fallback - Our Enhancement) =====
# Optional OCR support for scanned PDFs.
# Also install system tools: `tesseract-ocr` and `poppler-utils`.
pytesseract
pdf2image
Pillow
